Author level sprite quick stats

See merge request !231

Former-commit-id: aa34f545f9d6a50b417e7ded1a5149f2b962b1bf

+++ src/gameplayer/front_end/gui_generator/button_generator/IButton.java
@@ -6,6 +6,6 @@ import javafx.scene.input.MouseEvent;
 
 public interface IButton {
 	
-	public Button createButton(String aMessage, int aXPos, int aYPos, EventHandler<? super MouseEvent> aHandler);
+	public Button createButton(String aMessage, double aXPos, double aYPos, EventHandler<? super MouseEvent> aHandler);
 	
 }
